refactor: sortable virtual paginated table now has no unmatching UI error
